% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHealth.R
\name{checkHealth}
\alias{checkHealth}
\title{Check the Gobbler's health}
\usage{
checkHealth(staging)
}
\arguments{
\item{staging}{String containing the path to the staging directory.}
}
\value{
Logical scalar indicating whether the Gobbler responded within 10 seconds.
}
\description{
Check that the Gobbler is still active and monitoring the staging directory.
}
\examples{
if (interactive()) {
    refreshUsage("test-R")
}

}
\author{
Aaron Lun
}
