% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAssets.R
\name{listAssets}
\alias{listAssets}
\title{List assets}
\usage{
listAssets(project, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of asset names.
}
\description{
List all assets in a project.
}
\examples{
# Mocking up an upload. 
info <- startGobbler()
src <- allocateUploadDirectory(info$staging)
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
res <- uploadDirectory("test", "more-simple", "v1", src, staging=info$staging)
res <- uploadDirectory("test", "even-more-simple", "v1", src, staging=info$staging)

# Listing available assets:
listAssets("test", registry=info$registry)

}
\author{
Aaron Lun
}
