% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVersions.R
\name{listVersions}
\alias{listVersions}
\title{List asset versions}
\usage{
listVersions(project, asset, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of versions.
}
\description{
List all versions of a project asset.
}
\examples{
listVersions("test-R", "basic")

}
\author{
Aaron Lun
}
