% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindexVersion.R
\name{reindexVersion}
\alias{reindexVersion}
\title{Reindex a versioned asset}
\usage{
reindexVersion(project, asset, version, staging, url)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{asset}{String containing the asset to remove.}

\item{version}{String containing the version of the asset to remove.}

\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}
}
\value{
\code{NULL} is invisibly returned if the version was successfully reindexed.
}
\description{
Reindex a version of an asset of a project from the registry.
This regenerates all of the internal \code{..manifest} and \code{..links} files.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start with a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up a version if it doesn't already exist.
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
uploadDirectory("test", "simple", "v1", src, staging=info$staging, url=info$url)
fetchManifest("test", "simple", "v1", registry=info$registry, url=info$url)

# Let's add a new file directly to the directory.
write(file=file.path(info$registry, "test", "simple", "v1", "whee"), "stuff")

# And reindexing the version.
reindexVersion("test", "simple", "v1", staging=info$staging, url=info$url)
fetchManifest("test", "simple", "v1", registry=info$registry, url=info$url)

}
\author{
Aaron Lun
}
