% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeProject.R
\name{removeProject}
\alias{removeProject}
\title{Remove a project}
\usage{
removeProject(project, staging)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{staging}{String containing the path to the staging directory.}
}
\value{
\code{NULL} is invisibly returned if the project was successfully removed.
}
\description{
Remove a project from the registry.
}
\examples{
info <- startGobbler()

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging) 
listProjects("test", registry=info$registry)

# Removing the asset.
removeProject("test", staging=info$staging)
listProjects("test", registry=info$registry)

}
\seealso{
\code{\link{removeAsset}} and \code{\link{removeVersion}}, to remove assets and versions respectively.
}
\author{
Aaron Lun
}
