% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeVersion.R
\name{removeVersion}
\alias{removeVersion}
\title{Remove a versioned asset}
\usage{
removeVersion(project, asset, version, staging, url)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{asset}{String containing the asset to remove.}

\item{version}{String containing the version of the asset to remove.}

\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}
}
\value{
\code{NULL} is invisibly returned if the version was successfully removed.
}
\description{
Remove a version of an asset of a project from the registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start with a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up a version if it doesn't already exist.
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
uploadDirectory("test", "simple", "v1", src, staging=info$staging, url=info$url)
listVersions("test", "simple", registry=info$registry)

# Removing the version.
removeVersion("test", "simple", "v1", staging=info$staging, url=info$url)
listVersions("test", "simple", registry=info$registry)

}
\seealso{
\code{\link{removeProject}} and \code{\link{removeAsset}}, to remove a project or asset.
}
\author{
Aaron Lun
}
