% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceInfo.R
\name{serviceInfo}
\alias{serviceInfo}
\title{Gobbler service information}
\usage{
serviceInfo(url)
}
\arguments{
\item{url}{String containing the URL of the gobbler REST API.}
}
\value{
List containing the location of the staging and registry directories.
}
\description{
Get information about the Gobbler service, namely the locations of the staging directory and registry.
}
\examples{
info <- startGobbler()
serviceInfo(info$url)

}
\author{
Aaron Lun
}
