% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPermissions.R
\name{setPermissions}
\alias{setPermissions}
\title{Set project permissions}
\usage{
setPermissions(
  project,
  registry,
  staging,
  owners = NULL,
  uploaders = NULL,
  append = TRUE
)
}
\arguments{
\item{project}{String containing the project name.}

\item{registry}{String containing a path to the registry.}

\item{staging}{String containing a path to the staging directory.}

\item{owners}{Character vector containing the user IDs for owners of this project.
If \code{NULL}, no change is made to the existing owners of the project.}

\item{uploaders}{List specifying the authorized uploaders for this project.
See the \code{uploaders} field in the \code{\link{fetchPermissions}} return value for the expected format. 
If \code{NULL}, no change is made to the existing uploaders of the project.}

\item{append}{Logical scalar indicating whether \code{owners} and \code{uploaders} should be appended to the existing owners and uploaders, respectively, of the project.
If \code{FALSE}, the \code{owners} and \code{uploaders} are used to replace the existing values.}
}
\value{
\code{NULL} is invisibly returned upon successful setting of the permissions.
}
\description{
Set the owner and uploader permissions for a project.
}
\examples{
if (interactive()) {
    # Creating a project for demonstration purposes.
    createProject("test-R-perms", owners="LTLA")

    # Setting extra permissions on this project.
    setPermissions("test-R-perms",
        owners="jkanche", 
        uploaders=list(list(id="lawremi", until=Sys.time() + 1000))
    )
}

}
\seealso{
\code{\link{fetchPermissions}}, to fetch the permissions.
}
\author{
Aaron Lun
}
