% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGobbler.R
\name{startGobbler}
\alias{startGobbler}
\alias{stopGobbler}
\title{Start and stop a Gobbler service}
\usage{
startGobbler(staging = tempfile(), registry = tempfile(), wait = 1)

stopGobbler()
}
\arguments{
\item{staging}{String containing a path to a staging directory.
Ignored if the service is already running.}

\item{registry}{String containing a path to a registry.
Ignored if the service is already running.}

\item{wait}{Integer specifying the number of seconds to wait for service initialization.}
}
\value{
For \code{startGobbler}, a list indicating whether a new service was set up, plus the locations of the staging directory and registry. 

For \code{stopGobbler}, any existing service is shut down, and \code{NULL} is invisibly returned.
}
\description{
This sets up a single Gobbler service for the entire R session, is intended for examples and tests.
Real Gobbler deployments should operate outside of R.
}
\examples{
startGobbler()

}
\seealso{
\url{https://github.com/ArtifactDB/gobbler}, for source code and binaries to build and run a Gobbler service.
}
