% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpackPath.R
\name{unpackPath}
\alias{unpackPath}
\title{Unpack a path to its project-asset-version combination}
\usage{
unpackPath(path)
}
\arguments{
\item{path}{String containing a relative path within the Gobbler registry.}
}
\value{
List containing \code{project}, \code{asset}, \code{version} and \code{path}.
All are strings, except for \code{path}, which may be \code{NULL}.
}
\description{
Unpack a Gobbler path to its combination of project, asset, version, and (optionally) path,
for easier use in the various \pkg{gobbler} functions.
}
\examples{
unpackPath("project/asset/version/path")
unpackPath("project/asset/version")

}
\author{
Aaron Lun
}
