% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDirectory.R
\name{uploadDirectory}
\alias{uploadDirectory}
\title{Upload a directory's contents}
\usage{
uploadDirectory(
  project,
  asset,
  version,
  directory,
  staging,
  url,
  probation = FALSE
)
}
\arguments{
\item{project}{String containing an existing project name.}

\item{asset}{String containing a new or existing asset name in \code{project}.}

\item{version}{String containing the name of a new version of \code{asset}.}

\item{directory}{String containing the path to a directory to be uploaded.
This should be inside \code{staging} for best performance, typically using the directory allocated by \code{\link{allocateUploadDirectory}}.
Otherwise, the contents of \code{directory} will need to be copied to \code{staging} prior to upload.}

\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}

\item{probation}{Logical scalar indicating whether to upload a probational version.}
}
\value{
On success, \code{NULL} is invisibly returned.
}
\description{
Upload a directory as a new versioned asset of a project in the Gobbler registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start with a clean slate.
createProject("test", info$staging, url=info$url)

src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")

uploadDirectory(
    project="test", 
    asset="simple", 
    version="v1", 
    directory=src, 
    staging=info$staging,
    url=info$url
)

}
\seealso{
\code{\link{createProject}}, to create a new project for uploads.

\code{\link{versionPath}}, to obtain the path to a versioned asset's contents inside the registry.

\code{\link{fetchManifest}}, to obtain the manifest of the versioned asset's contents.
}
\author{
Aaron Lun
}
