% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionPath.R
\name{versionPath}
\alias{versionPath}
\title{Path to a versioned asset}
\usage{
versionPath(project, asset, version, ...)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{...}{Further arguments to pass to \code{\link{fetchDirectory}}.}
}
\value{
String containing a path to a versioned asset in the registry.
}
\description{
Obtain the path to the directory containing a versioned asset in the registry.
}
\examples{
info <- startGobbler()
versionPath("test", "simple", "v1", registry=info$registry, url=info$url)

}
\author{
Aaron Lun
}
