% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateUploadDirectory.R
\name{allocateUploadDirectory}
\alias{allocateUploadDirectory}
\title{Allocate a directory for upload}
\usage{
allocateUploadDirectory(staging, create = TRUE)
}
\arguments{
\item{staging}{String containing the path to the staging directory.}

\item{create}{Logical scalar specifying whether to actually create the subdirectory.}
}
\value{
String containing a path to a newly created subdirectory within \code{staging}.
This (or its subdirectories) are typically used as the \code{directory} argument in \code{\link{uploadDirectory}}.

If \code{create=FALSE}, a name will be chosen but the subdirectory itself will not yet be created.
}
\description{
Create a new subdirectory in the staging directory to store files for an upload to the Gobbler registry.
If the to-be-uploaded files are saved here, users can avoid an extra copying step in \code{\link{uploadDirectory}}.
}
\examples{
staging <- tempfile()
dir.create(staging)
allocateUploadDirectory(staging)

}
\author{
Aaron Lun
}
