% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateUploadDirectory.R
\name{allocateUploadDirectory}
\alias{allocateUploadDirectory}
\title{Allocate a directory for upload}
\usage{
allocateUploadDirectory(staging, attempts = 3)
}
\arguments{
\item{staging}{String containing the path to the staging directory.}

\item{attempts}{Integer scalar specifying the number of attempts to allocate a new directory, after which an error is thrown.}
}
\value{
String containing a path to a newly created subdirectory within \code{staging}.
This (or its subdirectories) are typically used as the \code{directory} argument in \code{\link{uploadDirectory}}.
}
\description{
Create a new subdirectory in the staging directory to store files for an upload to the Gobbler registry.
If the to-be-uploaded files are saved here, users can avoid an extra copying step in \code{\link{uploadDirectory}}.
}
\examples{
staging <- tempfile()
dir.create(staging)
allocateUploadDirectory(staging)

}
\author{
Aaron Lun
}
