% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approveProbation.R
\name{approveProbation}
\alias{approveProbation}
\title{Approve a probational upload}
\usage{
approveProbation(project, asset, version, staging, url, spoof = NULL)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}

\item{spoof}{String containing the name of a user on whose behalf this request is being made.
This should only be used if the Gobbler service allows spoofing by the current user.
If \code{NULL}, no spoofing is performed.}
}
\value{
\code{NULL} is invisibly returned upon successful approval.
}
\description{
Pretty much as it says: approve a probational upload of a version of a project's asset.
This removes the \code{on_probation} tag from the uploaded version.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # to start with a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up an upload.
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
uploadDirectory("test", "probation", "v1", src,
    staging=info$staging, url=info$url, probation=TRUE)
fetchSummary("test", "probation", "v1", registry=info$registry)

# After approval, the probation status disappears.
approveProbation("test", "probation", "v1", staging=info$staging, url=info$url)
fetchSummary("test", "probation", "v1", registry=info$registry)

}
\seealso{
\code{\link{rejectProbation}}, to reject the probational upload.

\code{\link{uploadDirectory}}, to specify probational uploads.
}
\author{
Aaron Lun
}
