% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHealth.R
\name{checkHealth}
\alias{checkHealth}
\title{Check the Gobbler's health}
\usage{
checkHealth(staging, url)
}
\arguments{
\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}
}
\value{
Logical scalar indicating whether the Gobbler is active.
}
\description{
Check that the Gobbler is still active and monitoring the staging directory.
}
\examples{
info <- startGobbler()
checkHealth(info$staging, info$url)

stopGobbler()
checkHealth(info$staging, info$url)

}
\author{
Aaron Lun
}
