% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Create a project}
\usage{
createProject(project, staging, owners = NULL, uploaders = NULL)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{staging}{String containing the path to the staging directory.}

\item{owners}{Character vector containing the user IDs for owners of this project.
This defaults to the current user.}

\item{uploaders}{List specifying the authorized uploaders for this project.
See the \code{uploaders} field in the \code{\link{fetchPermissions}} return value for the expected format.}
}
\value{
\code{NULL} is invisibly returned if the project was successfully created.
}
\description{
Create a new project in the registry.
}
\examples{
info <- startGobbler()
removeProject("test", staging=info$staging) # start with a clean slate.

# Creating our new project.
createProject("test", staging=info$staging)
listProjects(registry=info$registry)

}
\seealso{
\code{\link{uploadDirectory}}, to upload a new version of an asset to an existing project.

\code{\link{removeProject}}, to remove a project respectively.
}
\author{
Aaron Lun
}
