% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSummary.R
\name{fetchSummary}
\alias{fetchSummary}
\title{Fetch version summary}
\usage{
fetchSummary(project, asset, version, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{registry}{String containing the path to the registry.}
}
\value{
List containing the summary for this version, with the following fields:
\itemize{
\item \code{upload_user_id}, string containing the identity of the uploader.
\item \code{upload_start}, a \link{POSIXct} object containing the upload start time.
\item \code{upload_finish}, a \link{POSIXct} object containing the upload finish time.
\item \code{on_probation} (optional), a logical scalar indicating whether the upload is probational.
If missing, this can be assumed to be \code{FALSE}.
}
}
\description{
Fetch the summary for a version of an asset of a project.
}
\examples{
# Mocking up an upload. 
info <- startGobbler()
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
stopGobbler(info, keep.dir=TRUE)

# Obtain a summary for this version.
fetchSummary("test", "simple", "v1", registry=info$registry)

}
\author{
Aaron Lun
}
