% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchUsage.R
\name{fetchUsage}
\alias{fetchUsage}
\title{Fetch project usage details}
\usage{
fetchUsage(project, registry, url, forceRemote = FALSE)
}
\arguments{
\item{project}{String containing the project name.}

\item{registry}{String containing a path to the registry.}

\item{url}{String containing the URL to the Gobbler REST API.}

\item{forceRemote}{Logical scalar indicating whether to force remote access via the API,
even if \code{registry} is accessible to the caller.}
}
\value{
Numeric scalar specifying the quota usage for the project, in bytes.
}
\description{
Fetch the quota usage for a project.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start with a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
write(file=file.path(src, "whee"), "stuff")
uploadDirectory("test", "simple", "v1", src, staging=info$staging, url=info$url)

# Obtaining the project usage.
fetchUsage("test", registry=info$registry)

}
\seealso{
\code{\link{refreshUsage}}, to recompute the used quota.
}
\author{
Aaron Lun
}
