% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAssets.R
\name{listAssets}
\alias{listAssets}
\title{List assets}
\usage{
listAssets(project, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of asset names.
}
\description{
List all assets in a project.
}
\examples{
info <- startGobbler()

# Mocking up a few uploads.
src <- allocateUploadDirectory(info$staging)
removeProject("test", info$staging) # clean out existing entries
for (ass in c("simple", "more-simple", "even-more-simple")) {
    uploadDirectory("test", ass, "v1", src, staging=info$staging)
}

# Listing available assets:
listAssets("test", registry=info$registry)

}
\author{
Aaron Lun
}
