% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAssets.R
\name{listAssets}
\alias{listAssets}
\title{List assets}
\usage{
listAssets(project, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of asset names.
}
\description{
List all assets in a project.
}
\examples{
listAssets("test-R")

}
\author{
Aaron Lun
}
