% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{List files for a version}
\usage{
listFiles(project, asset, version, registry, prefix = NULL, include.. = TRUE)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{registry}{String containing a path to the registry.}

\item{prefix}{String containing the prefix for the path.
If provided, files are only listed if they have a relative path (i.e., inside the version subdirectory) that starts with this prefix.
If \code{NULL}, all files associated with this version are listed.}

\item{include..}{Logical scalar indicating whether to list files with path components that start with \code{..}.}
}
\value{
Character vector of relative paths of files associated with the versioned asset.
}
\description{
List the contents of a version of a project asset.
}
\examples{
# Mocking up an upload. 
info <- startGobbler()
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
dir.create(file.path(src, "whee"))
write(file=file.path(src, "whee", "blah"), "stuff")
write(file=file.path(src, "whee2"), "more-stuff")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)

# List files, with or without a prefix.
listFiles("test", "simple", "v1", registry=info$registry)
listFiles("test", "simple", "v1", registry=info$registry, prefix="whee")
listFiles("test", "simple", "v1", registry=info$registry, prefix="whee/")

}
\author{
Aaron Lun
}
