% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(registry)
}
\arguments{
\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the registry.
}
\examples{
# Mocking up an upload. 
info <- startGobbler()
src <- allocateUploadDirectory(info$staging)
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
res <- uploadDirectory("more-tests", "simple", "v1", src, staging=info$staging)
res <- uploadDirectory("even-more-tests", "simple", "v1", src, staging=info$staging)

# Listing out the projects.
listProjects(info$registry)

}
\author{
Aaron Lun
}
