% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(registry, url, forceRemote = FALSE)
}
\arguments{
\item{registry}{String containing a path to the registry.}

\item{url}{String containing the URL to the Gobbler REST API.
Only used for remote access.}

\item{forceRemote}{Logical scalar indicating whether to force remote access via the API,
even if \code{registry} is on the same filesystem as the caller.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the registry.
This will call the REST API if the caller is not on the same filesystem as the registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # clean out any existing entries
removeProject("more-test", info$staging, url=info$url)
removeProject("even-more-test", info$staging, url=info$url)

# Now mocking up the creation of some projects.
createProject("test", info$staging, url=info$url)
createProject("more-test", info$staging, url=info$url)
createProject("even-more-test", info$staging, url=info$url)

# Listing out the projects.
listProjects(info$registry, info$url)

# Forcing remote access.
listProjects(info$registry, info$url, forceRemote=TRUE)

}
\author{
Aaron Lun
}
