% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(registry)
}
\arguments{
\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the registry.
}
\examples{
info <- startGobbler()

# Mocking up a few uploads.
src <- allocateUploadDirectory(info$staging)
for (proj in c("test", "more-tests", "even-more-tests")) {
    removeProject(proj, info$staging) # clean out any existing entry
    uploadDirectory(proj, "simple", "v1", src, staging=info$staging)
}

# Listing out the projects.
listProjects(info$registry)

}
\author{
Aaron Lun
}
