% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(registry)
}
\arguments{
\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging) # clean out any existing entries
removeProject("more-test", info$staging)
removeProject("even-more-test", info$staging)
   
# Now mocking up the creation of some projects.
createProject("test", info$staging)
createProject("more-test", info$staging)
createProject("even-more-test", info$staging)

# Listing out the projects.
listProjects(info$registry)

}
\author{
Aaron Lun
}
