% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all projects}
\usage{
listProjects(registry)
}
\arguments{
\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of project names.
}
\description{
List all projects in the registry.
}
\examples{
if (interactive()) {
    listProjects()
}

}
\author{
Aaron Lun
}
