% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVersions.R
\name{listVersions}
\alias{listVersions}
\title{List asset versions}
\usage{
listVersions(project, asset, registry, url, forceRemote = FALSE)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{registry}{String containing a path to the registry.}

\item{url}{String containing the URL to the Gobbler REST API.
Only used for remote access.}

\item{forceRemote}{Logical scalar indicating whether to force remote access via the API,
even if \code{registry} is on the same filesystem as the caller.}
}
\value{
Character vector of versions.
}
\description{
List all versions of a project asset.
This will call the REST API if the caller is not on the same filesystem as the registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start from a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up a few uploads.
src <- allocateUploadDirectory(info$staging)
for (v in c("v1", "v2")) {
    uploadDirectory("test", "simple", v, src, staging=info$staging, url=info$url)
}

# Listing the versions of the asset:
listVersions("test", "simple", registry=info$registry, url=info$url)

# Force remote listing:
listVersions("test", "simple", registry=info$registry, url=info$url, forceRemote=TRUE)
}
\author{
Aaron Lun
}
