% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVersions.R
\name{listVersions}
\alias{listVersions}
\title{List asset versions}
\usage{
listVersions(project, asset, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{registry}{String containing a path to the registry.}
}
\value{
Character vector of versions.
}
\description{
List all versions of a project asset.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging) # start from a clean slate.
createProject("test", info$staging) 

# Mocking up a few uploads.
src <- allocateUploadDirectory(info$staging)
for (v in c("v1", "v2")) {
    uploadDirectory("test", "simple", v, src, staging=info$staging)
}

# Listing the versions of the asset:
listVersions("test", "simple", registry=info$registry)

}
\author{
Aaron Lun
}
