% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshLatest.R
\name{refreshLatest}
\alias{refreshLatest}
\title{Refresh the latest version}
\usage{
refreshLatest(project, asset, staging)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{staging}{String containing a path to the staging directory.}
}
\value{
String containing the latest version of the project, or \code{NULL} if there are no non-probational versions.
}
\description{
Recompute the latest version of a project's asset.
This is useful on rare occasions where multiple simultaneous uploads cause the latest version to be slightly out of sync.
}
\examples{
if (interactive()) {
    refreshLatest("test-R", "basic")
}

}
\seealso{
\code{\link{fetchLatest}}, to get the latest version without recomputing it.
}
\author{
Aaron Lun
}
