% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshUsage.R
\name{refreshUsage}
\alias{refreshUsage}
\title{Refresh the quota usage}
\usage{
refreshUsage(project, staging, url)
}
\arguments{
\item{project}{String containing the project name.}

\item{staging}{String containing the path to the staging directory.}

\item{url}{String containing the URL of the gobbler REST API.}
}
\value{
Numeric scalar specifying the total quota usage of this project, in bytes.
}
\description{
Recompute the quota usage of a project.
This is useful on rare occasions where multiple simultaneous uploads cause the usage calculations to be out of sync.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging, url=info$url) # start with a clean slate.
createProject("test", info$staging, url=info$url)

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
write(file=file.path(src, "whee"), "stuff")
uploadDirectory("test", "simple", "v1", src, staging=info$staging, url=info$url)

# Messing with the project usage:
write(file=file.path(info$registry, "test", "..usage"), '{ "total": 0 }')
fetchUsage("test", registry=info$registry)

# Fixing the project usage.
refreshUsage("test", staging=info$staging, url=info$url)
fetchUsage("test", registry=info$registry)

}
\seealso{
\code{\link{fetchUsage}}, to get the usage without recomputing it.
}
\author{
Aaron Lun
}
