% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshUsage.R
\name{refreshUsage}
\alias{refreshUsage}
\title{Refresh the quota usage}
\usage{
refreshUsage(project, staging)
}
\arguments{
\item{project}{String containing the project name.}

\item{staging}{String containing the path to the staging directory.}
}
\value{
Numeric scalar specifying the total quota usage of this project, in bytes.
}
\description{
Recompute the quota usage of a project.
This is useful on rare occasions where multiple simultaneous uploads cause the usage calculations to be out of sync.
}
\examples{
info <- startGobbler()

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
write(file=file.path(src, "whee"), "stuff")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)

# Messing with the project usage:
write(file=file.path(info$registry, "test", "..usage"), '{ "total": 0 }')
fetchUsage("test", registry=info$registry)

# Fixing the project usage.
refreshUsage("test", staging=info$staging)
fetchUsage("test", registry=info$registry)

stopGobbler(info, keep.dir=TRUE)
}
\seealso{
\code{\link{fetchUsage}}, to get the usage without recomputing it.
}
\author{
Aaron Lun
}
