% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejectProbation.R
\name{rejectProbation}
\alias{rejectProbation}
\title{Reject a probational upload}
\usage{
rejectProbation(project, asset, version, staging)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{staging}{String containing the path to the staging directory.}
}
\value{
\code{NULL} is invisibly returned upon successful rejection.
}
\description{
Pretty much as it says: reject a probational upload of a version of a project's asset.
This removes all files associated with that version.
}
\examples{
info <- startGobbler()

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "probation", "v1", src, staging=info$staging, probation=TRUE)
listVersions("test", "probation", registry=info$registry)

# After rejection, the version disppears.
rejectProbation("test", "probation", "v1", staging=info$staging)
listVersions("test", "probation", registry=info$registry)

stopGobbler(info)

}
\seealso{
\code{\link{approveProbation}}, to approve the probational upload.

\code{\link{uploadDirectory}}, to specify probational uploads.
}
\author{
Aaron Lun
}
