% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAsset.R
\name{removeAsset}
\alias{removeAsset}
\title{Remove an asset of a project}
\usage{
removeAsset(project, asset, staging)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{asset}{String containing the asset to remove.}

\item{staging}{String containing the path to a staging directory.}
}
\value{
\code{NULL} is invisibly returned if the asset was successfully removed.
}
\description{
Remove an asset of the project from the registry.
}
\examples{
info <- startGobbler()

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
listAssets("test", registry=info$registry)

# Removing the asset.
removeAsset("test", "simple", staging=info$staging)
listAssets("test", registry=info$registry)

}
\seealso{
\code{\link{removeProject}} and \code{\link{removeVersion}}, to remove projects and versions respectively.
}
\author{
Aaron Lun
}
