% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeProject.R
\name{removeProject}
\alias{removeProject}
\title{Remove a project}
\usage{
removeProject(project, staging)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{staging}{String containing the path to the staging directory.}
}
\value{
\code{NULL} is invisibly returned if the project was successfully removed.
}
\description{
Remove a project from the registry.
}
\examples{
info <- startGobbler()
removeProject("test", info$staging) # start with a clean slate.

# Mocking up a project so we have something to delete.
createProject("test", info$staging)
listProjects(registry=info$registry)

# Removing the project.
removeProject("test", staging=info$staging)
listProjects(registry=info$registry)

}
\seealso{
\code{\link{removeAsset}} and \code{\link{removeVersion}}, to remove assets and versions respectively.

\code{\link{createProject}}, to create a project.
}
\author{
Aaron Lun
}
