% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeVersion.R
\name{removeVersion}
\alias{removeVersion}
\title{Remove a versioned asset}
\usage{
removeVersion(project, asset, version, staging)
}
\arguments{
\item{project}{String containing the project to remove.}

\item{asset}{String containing the asset to remove.}

\item{version}{String containing the version of the asset to remove.}

\item{staging}{String containing a path to a staging directory.}
}
\value{
\code{NULL} is invisibly returned if the version was successfully removed.
}
\description{
Remove a version of an asset of a project from the registry.
}
\examples{
info <- startGobbler()

# Mocking up an upload. 
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
listVersions("test", "simple", registry=info$registry)

# Removing the version.
removeVersion("test", "simple", "v1", staging=info$staging)
listProjects("test", "simple", registry=info$registry)

}
\seealso{
\code{\link{removeProject}} and \code{\link{removeAsset}}, to remove a project or asset.
}
\author{
Aaron Lun
}
