% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGobbler.R
\name{startGobbler}
\alias{startGobbler}
\alias{stopGobbler}
\title{Start and stop a Gobbler service}
\usage{
startGobbler(staging = tempfile(), registry = tempfile())

stopGobbler(info, keep.dir = FALSE)
}
\arguments{
\item{staging}{String containing a path to a staging directory.}

\item{registry}{String containing a path to a registry.}

\item{info}{List created by \code{startGobbler}.}

\item{keep.dir}{Logical scalar indicating whether the staging directory and regisry should be retained.}
}
\value{
\code{startGobbler} starts a Gobbler service and returns a list including the locations of the staging directory and registry. 

\code{stopGobbler} stops the Gobbler service, deletes the relevant directories if \code{keep.dir=FALSE}, and returns \code{NULL} invisibly.
}
\description{
This is intended for examples and tests.
Real Gobbler deployments should operate outside of R.
}
\examples{
info <- startGobbler()
stopGobbler(info)

}
\seealso{
\url{https://github.com/ArtifactDB/gobbler}, for source code and binaries to build and run a Gobbler service.
}
