% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGobbler.R
\name{startGobbler}
\alias{startGobbler}
\alias{stopGobbler}
\title{Start and stop a Gobbler service}
\usage{
startGobbler(
  staging = tempfile(),
  registry = tempfile(),
  port = NULL,
  wait = 1,
  version = "0.3.4",
  overwrite = FALSE
)

stopGobbler()
}
\arguments{
\item{staging}{String containing a path to a staging directory.
Ignored if the service is already running.}

\item{registry}{String containing a path to a registry.
Ignored if the service is already running.}

\item{port}{Integer specifying the port to use for hosting the service.
If \code{NULL}, a free port is randomly selected.}

\item{wait}{Integer specifying the number of seconds to wait for service initialization.}

\item{version}{String containing the desired version of the Gobbler binary.}

\item{overwrite}{Logical scalar indicating whether to redownload the Gobbler binary.}
}
\value{
For \code{startGobbler}, a list containing:
\itemize{
\item \code{new}, a logical scalar indicating whether a new service was set up.
\item \code{staging}, string containing the location of the staging directory.
\item \code{registry}, string containing the location of the registry.
\item \code{port}, integer containing the port.
\item \code{url}, string containing the URL of the REST API.
}

For \code{stopGobbler}, any existing service is shut down, and \code{NULL} is invisibly returned.
}
\description{
This sets up a single Gobbler service for the entire R session, is intended for examples and tests.
Real Gobbler deployments should operate outside of R.
}
\examples{
startGobbler()
startGobbler() # repeated calls just re-use the same instance.

stopGobbler()
startGobbler() # reinitializes a new instance.

}
\seealso{
\url{https://github.com/ArtifactDB/gobbler}, for source code and binaries to build and run a Gobbler service.
}
