% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDirectory.R
\name{uploadDirectory}
\alias{uploadDirectory}
\title{Upload a directory's contents}
\usage{
uploadDirectory(
  project,
  asset,
  version,
  dir,
  staging,
  prefix = NULL,
  probation = FALSE,
  owners = NULL,
  uploaders = NULL
)
}
\arguments{
\item{project}{String containing the project name.
This may be \code{NULL}, in which case \code{prefix} should be provided.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.
This may be \code{NULL} to generate a new version name based on an incrementing series.}

\item{dir}{String containing the path to a directory to be uploaded.
This should be inside \code{staging} for best performance, see \code{\link{allocateUploadDirectory}} for details.}

\item{staging}{String containing the path to the staging directory.}

\item{prefix}{String containing the project name prefix.
This is used to derive a project name based on an incrementing series.}

\item{probation}{Logical scalar indicating whether to upload a probational version.}

\item{owners}{Character vector containing the user IDs for owners of this project.
This defaults to the current user.}

\item{uploaders}{List specifying the authorized uploaders for this project.
See the \code{uploaders} field in the \code{\link{fetchPermissions}} return value for the expected format.}
}
\value{
List containing \code{project} and \code{version},
the generated names for the project and version respectively.
(If these were explicitly provided in the arguments, the same values are returned verbatim.)
}
\description{
Upload a directory as a new versioned asset of a project in the registry.
}
\examples{
info <- startGobbler()

src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory(
    project=NULL, 
    prefix="TEST", 
    asset="simple", 
    version=NULL, 
    dir=src, 
    staging=info$staging)
res

}
\seealso{
\code{\link{versionPath}}, to obtain the path to a versioned asset's contents inside the registry.

\code{\link{listManifest}}, to obtain the manifest of the versioned asset's contents.
}
\author{
Aaron Lun
}
