% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versionPath.R
\name{versionPath}
\alias{versionPath}
\title{Path to a versioned asset}
\usage{
versionPath(project, asset, version, registry)
}
\arguments{
\item{project}{String containing the project name.}

\item{asset}{String containing the asset name.}

\item{version}{String containing the version name.}

\item{registry}{String containing the path to the registry.}
}
\value{
String containing a path to a versioned asset in the registry.
}
\description{
Obtain the path to the directory containing a versioned asset in the registry.
}
\examples{
# Mocking up an upload. 
info <- startGobbler()
src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")
res <- uploadDirectory("test", "simple", "v1", src, staging=info$staging)
stopGobbler(info, keep.dir=TRUE)

# Obtaining the path.
versionPath("test", "simple", "v1", registry=info$registry)

}
\author{
Aaron Lun
}
